+++ context/app/static/js/api/scfind/scfind.test.ts
@@ -1,4 +1,4 @@
-import { createScFindKey } from './utils';
+import { createScFindKey, extractCellTypeInfo, extractCellTypesInfo, formatCellTypeName } from './utils';
 
 const scFindEndpoint = 'http://example.com';
 
@@ -41,5 +41,84 @@ describe('createScfindKey', () => {
       expect(key.includes(k)).toBe(false);
     });
     expectURLIsValid(key);
+  });
+});
+
+describe('formatCellTypeName', () => {
+  it('should remove organ prefix from cell type name', () => {
+    const cellType = 'organ.cellType';
+    const formatted = formatCellTypeName(cellType);
+    expect(formatted).toBe('cellType');
+  });
+
+  it('should return the same name if no organ prefix is present', () => {
+    const cellType = 'cellType';
+    const formatted = formatCellTypeName(cellType);
+    expect(formatted).toBe('cellType');
+  });
+
+  it('should handle empty strings', () => {
+    const cellType = '';
+    const formatted = formatCellTypeName(cellType);
+    expect(formatted).toBe('');
+  });
+});
+
+describe('extractCellTypeInfo', () => {
+  it('should extract organ, name, and variant from cell type string', () => {
+    const cellType = 'brain.neuron:adult';
+    const result = extractCellTypeInfo(cellType);
+    expect(result).toEqual({ organ: 'brain', name: 'neuron', variant: 'adult' });
+  });
+
+  it('should handle cell types without variants', () => {
+    const cellType = 'heart.muscle';
+    const result = extractCellTypeInfo(cellType);
+    expect(result).toEqual({ organ: 'heart', name: 'muscle', variant: undefined });
+  });
+
+  it('should handle empty cell type string', () => {
+    const cellType = '';
+    const result = extractCellTypeInfo(cellType);
+    expect(result).toEqual({ organ: '', name: '', variant: undefined });
+  });
+});
+
+describe('extractCellTypesInfo', () => {
+  it('should extract name, organs, and variants from cell types array', () => {
+    const cellTypes = ['brain.neuron:adult', 'heart.muscle', 'lung.epithelial:child'];
+    const result = extractCellTypesInfo(cellTypes);
+    expect(result).toEqual({
+      name: 'neuron',
+      organs: ['brain', 'heart', 'lung'],
+      variants: {
+        brain: ['adult'],
+        heart: [],
+        lung: ['child'],
+      },
+    });
+  });
+
+  it('should handle empty cell types array', () => {
+    const cellTypes: string[] = [];
+    const result = extractCellTypesInfo(cellTypes);
+    expect(result).toEqual({
+      name: '',
+      organs: [],
+      variants: {},
+    });
+  });
+
+  it('should handle cell types with no variants', () => {
+    const cellTypes = ['brain.neuron', 'heart.muscle'];
+    const result = extractCellTypesInfo(cellTypes);
+    expect(result).toEqual({
+      name: 'neuron',
+      organs: ['brain', 'heart'],
+      variants: {
+        brain: [],
+        heart: [],
+      },
+    });
   });
 });
