+++ context/app/static/js/helpers/is-error.ts
@@ -1,3 +1,3 @@
-export function objectIsError(response: object): response is Error {
+export function isError(response: object): response is Error {
   return 'message' in response;
 }
